+++ src-tauri/src/db/schema.rs
@@ -13,7 +13,7 @@ diesel::table! {
         user_id -> Text,
         game_id -> Text,
         value_name -> Text,
-        value_num -> BigInt,
+        value_num -> Double,
     }
 }
